Core/DBLayer: Support using mysql 8
(cherry picked from commit c3a9d56b56b665133707f587ecb1bd1c272f6911)
