+++ cmake/macros/FindMySQL.cmake
@@ -86,19 +86,25 @@ find_path(MYSQL_INCLUDE_DIR
     /usr/local/include
     /usr/local/include/mysql
     /usr/local/mysql/include
+    "${PROGRAM_FILES_64}/MySQL/MySQL Server 8.0/include"
     "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.7/include"
     "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.6/include"
     "${PROGRAM_FILES_64}/MySQL/include"
+    "${PROGRAM_FILES_32}/MySQL/MySQL Server 8.0/include"
     "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.7/include"
     "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.6/include"
     "${PROGRAM_FILES_32}/MySQL/include"
     "C:/MySQL/include"
+    "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 8.0;Location]/include"
     "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.7;Location]/include"
     "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.6;Location]/include"
+    "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 8.0;Location]/include"
     "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.7;Location]/include"
     "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.6;Location]/include"
+    "$ENV{ProgramFiles}/MySQL/MySQL Server 8.0/include"
     "$ENV{ProgramFiles}/MySQL/MySQL Server 5.7/include"
     "$ENV{ProgramFiles}/MySQL/MySQL Server 5.6/include"
+    "$ENV{SystemDrive}/MySQL/MySQL Server 8.0/include"
     "$ENV{SystemDrive}/MySQL/MySQL Server 5.7/include"
     "$ENV{SystemDrive}/MySQL/MySQL Server 5.6/include"
     "c:/msys/local/include"
@@ -130,27 +136,37 @@ if( WIN32 )
       libmysql
     PATHS
       ${MYSQL_ADD_LIBRARIES_PATH}
+      "${PROGRAM_FILES_64}/MySQL/MySQL Server 8.0/lib"
       "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.7/lib"
       "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.6/lib"
+      "${PROGRAM_FILES_64}/MySQL/MySQL Server 8.0/lib/opt"
       "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.7/lib/opt"
       "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.6/lib/opt"
       "${PROGRAM_FILES_64}/MySQL/lib"
+      "${PROGRAM_FILES_32}/MySQL/MySQL Server 8.0/lib"
       "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.7/lib"
       "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.6/lib"
+      "${PROGRAM_FILES_32}/MySQL/MySQL Server 8.0/lib/opt"
       "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.7/lib/opt"
       "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.6/lib/opt"
       "${PROGRAM_FILES_32}/MySQL/lib"
       "C:/MySQL/lib/debug"
+      "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 8.0;Location]/lib"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.7;Location]/lib"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.6;Location]/lib"
+      "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 8.0;Location]/lib/opt"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.7;Location]/lib/opt"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.6;Location]/lib/opt"
+      "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 8.0;Location]/lib"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.7;Location]/lib"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.6;Location]/lib"
+      "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 8.0;Location]/lib/opt"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.7;Location]/lib/opt"
       "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.6;Location]/lib/opt"
+      "$ENV{ProgramFiles}/MySQL/MySQL Server 8.0/lib/opt"
       "$ENV{ProgramFiles}/MySQL/MySQL Server 5.7/lib/opt"
       "$ENV{ProgramFiles}/MySQL/MySQL Server 5.6/lib/opt"
+      "$ENV{SystemDrive}/MySQL/MySQL Server 8.0/lib/opt"
       "$ENV{SystemDrive}/MySQL/MySQL Server 5.7/lib/opt"
       "$ENV{SystemDrive}/MySQL/MySQL Server 5.6/lib/opt"
       "c:/msys/local/include"
@@ -191,27 +207,37 @@ endif( UNIX )
 if( WIN32 )
     find_program(MYSQL_EXECUTABLE mysql
       PATHS
+        "${PROGRAM_FILES_64}/MySQL/MySQL Server 8.0/bin"
         "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.7/bin"
         "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.6/bin"
+        "${PROGRAM_FILES_64}/MySQL/MySQL Server 8.0/bin/opt"
         "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.7/bin/opt"
         "${PROGRAM_FILES_64}/MySQL/MySQL Server 5.6/bin/opt"
         "${PROGRAM_FILES_64}/MySQL/bin"
+        "${PROGRAM_FILES_32}/MySQL/MySQL Server 8.0/bin"
         "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.7/bin"
         "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.6/bin"
+        "${PROGRAM_FILES_32}/MySQL/MySQL Server 8.0/bin/opt"
         "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.7/bin/opt"
         "${PROGRAM_FILES_32}/MySQL/MySQL Server 5.6/bin/opt"
         "${PROGRAM_FILES_32}/MySQL/bin"
         "C:/MySQL/bin/debug"
+        "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 8.0;Location]/bin"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.7;Location]/bin"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.6;Location]/bin"
+        "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 8.0;Location]/bin/opt"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.7;Location]/bin/opt"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\MySQL AB\\MySQL Server 5.6;Location]/bin/opt"
+        "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 8.0;Location]/bin"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.7;Location]/bin"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.6;Location]/bin"
+        "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 8.0;Location]/bin/opt"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.7;Location]/bin/opt"
         "[HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\MySQL AB\\MySQL Server 5.6;Location]/bin/opt"
+        "$ENV{ProgramFiles}/MySQL/MySQL Server 8.0/bin/opt"
         "$ENV{ProgramFiles}/MySQL/MySQL Server 5.7/bin/opt"
         "$ENV{ProgramFiles}/MySQL/MySQL Server 5.6/bin/opt"
+        "$ENV{SystemDrive}/MySQL/MySQL Server 8.0/bin/opt"
         "$ENV{SystemDrive}/MySQL/MySQL Server 5.7/bin/opt"
         "$ENV{SystemDrive}/MySQL/MySQL Server 5.6/bin/opt"
         "c:/msys/local/include"
