+++ src/server/database/CMakeLists.txt
@@ -50,9 +50,9 @@ add_definitions(-DTRINITY_API_EXPORT_DATABASE)
 target_link_libraries(database
   PRIVATE
     trinity-core-interface
+    mysql
   PUBLIC
-    common
-    mysql)
+    common)
 
 set_target_properties(database
     PROPERTIES
