+++ src/server/database/Database/Field.cpp
@@ -304,10 +304,7 @@ bool Field::IsNumeric() const
 
 #ifdef TRINITY_STRICT_DATABASE_TYPE_CHECKS
 
-#ifdef _WIN32 // hack for broken mysql.h not including the correct winsock header for SOCKET definition, fixed in 5.7
-#include <winsock2.h>
-#endif
-#include <mysql.h>
+#include "MySQLHacks.h"
 
 static char const* FieldTypeToString(enum_field_types type)
 {
@@ -343,7 +340,7 @@ static char const* FieldTypeToString(enum_field_types type)
     }
 }
 
-void Field::SetMetadata(MYSQL_FIELD* field, uint32 fieldIndex)
+void Field::SetMetadata(MySQLField* field, uint32 fieldIndex)
 {
     meta.TableName = field->org_table;
     meta.TableAlias = field->table;
